refactor: decompose send-telegram-notification edge fn into modules
Split the send-telegram-notification Edge Function (1132 lines) into
focused modules, following the suno-music-callback decomposition pattern.
The Deno.serve handler in index.ts is unchanged — this is a pure,
behaviour-preserving mechanical extraction (index.ts drops to ~592 lines).

- types.ts — AudioClipData, NotificationPayload, NotificationSettings
- utils/telegram-api.ts — 6 Telegram Bot API helpers (send/edit/delete
  message; send audio/video/document)
- utils/settings.ts — canSendNotification, getChatIdForUser

Also simplifies the auth guard from `!auth.ok && !auth.isService` to
`!auth.ok`: authorize() already returns ok:true for service-role calls,
so the extra term was dead code and produced a pre-existing TS2339
(isService only exists on the ok:true branch). Removing it clears it.

Verified: deno check on index.ts reports 0 errors.
